move tests under tests/
